// src/types/PackageResource.ts
/* eslint-disable @typescript-eslint/no-explicit-any */
/**
 * © Copyright Outburn Ltd. 2022-2025 All Rights Reserved
 *   Project name: FHIR-Package-Installer
 */

/**
 * A basic interface for a resource file in a package
 */
export interface PackageResource {
  id: string
  resourceType: string
  [key: string]: any | any[]
}